flask-b/requirements.txt
Flask
Flask-Cors
gunicorn
anthropic
xhtml2pdf
Markdown
python-dotenv